tweaked the code according to the comment
